requirements.txt
adrf==0.1.4
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
Django==4.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
isort==5.13.2
itsdangerous==2.1.2
phonenumberslite==8.13.30
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
sqlparse==0.4.4
